# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/email/activate": {
            "get": {
                "description": "activate email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "E-mail"
                ],
                "summary": "activate email",
                "parameters": [
                    {
                        "type": "string",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200, \"msg\":ok}",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseContent"
                        }
                    },
                    "400": {
                        "description": "40022:validation failed , 40102: token expired, 40101: unknown token invalid error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseContent"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/email/register": {
            "post": {
                "description": "Using email to register an account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "E-mail"
                ],
                "summary": "Register an account by email",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers_v1.RegisterByMailInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200, \"msg\":ok}",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseContent"
                        }
                    },
                    "400": {
                        "description": "40022:validation failed, 400009: already registered",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseContent"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/email/verify": {
            "post": {
                "description": "Using email to receive a login token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "E-mail"
                ],
                "summary": "Get login token by email",
                "parameters": [
                    {
                        "description": "login data",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers_v1.VerifyMailLoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200, \"msg\":ok}",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseContent"
                        }
                    },
                    "400": {
                        "description": "40004:user not found, 40000: email not verified yet",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseContent"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200, \"msg\":ok}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseContent"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "40000: invalid token",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseContent"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers_v1.RegisterByMailInput": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "dean.test@gmail.com"
                },
                "name": {
                    "type": "string",
                    "example": "dean"
                },
                "password": {
                    "type": "string",
                    "example": "!AS$GK())"
                }
            }
        },
        "handlers_v1.VerifyMailLoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "dean.test@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "!AS$GK())"
                }
            }
        },
        "helpers.ResponseContent": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "object"
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}